+++ subsys/tracing/ctf/ctf_top.c
@@ -443,7 +443,7 @@ void sys_trace_socket_accept_enter(int sock)
 }
 
 void sys_trace_socket_accept_exit(int sock, const struct sockaddr *addr,
-				  const size_t *addrlen, int ret)
+				  const uint32_t *addrlen, int ret)
 {
 	ctf_net_bounded_string_t addr_str = { "unknown" };
 	uint32_t addr_len = 0U;
@@ -463,7 +463,7 @@ void sys_trace_socket_accept_exit(int sock, const struct sockaddr *addr,
 }
 
 void sys_trace_socket_sendto_enter(int sock, int len, int flags,
-				   const struct sockaddr *dest_addr, size_t addrlen)
+				   const struct sockaddr *dest_addr, uint32_t addrlen)
 {
 	ctf_net_bounded_string_t addr_str = { "unknown" };
 
@@ -504,7 +504,7 @@ void sys_trace_socket_sendmsg_exit(int sock, int ret)
 }
 
 void sys_trace_socket_recvfrom_enter(int sock, int max_len, int flags,
-				     struct sockaddr *addr, size_t *addrlen)
+				     struct sockaddr *addr, uint32_t *addrlen)
 {
 	ctf_top_socket_recvfrom_enter(sock, max_len, flags,
 				      (uint32_t)(uintptr_t)addr,
@@ -512,7 +512,7 @@ void sys_trace_socket_recvfrom_enter(int sock, int max_len, int flags,
 }
 
 void sys_trace_socket_recvfrom_exit(int sock, const struct sockaddr *src_addr,
-				    const size_t *addrlen, int ret)
+				    const uint32_t *addrlen, int ret)
 {
 	ctf_net_bounded_string_t addr_str = { "unknown" };
 	int len = 0;
@@ -631,7 +631,7 @@ void sys_trace_socket_getpeername_enter(int sock)
 }
 
 void sys_trace_socket_getpeername_exit(int sock,  struct sockaddr *addr,
-				       const size_t *addrlen, int ret)
+				       const uint32_t *addrlen, int ret)
 {
 	ctf_net_bounded_string_t addr_str;
 
@@ -647,7 +647,7 @@ void sys_trace_socket_getsockname_enter(int sock)
 }
 
 void sys_trace_socket_getsockname_exit(int sock, const struct sockaddr *addr,
-				       const size_t *addrlen, int ret)
+				       const uint32_t *addrlen, int ret)
 {
 	ctf_net_bounded_string_t addr_str;
